v1 falta arreglar seleccion de imagen de galeria

+++ app/src/main/java/com/practica/ejemplodagger/sis/viewmodel/CategoriaViewModel.kt
@@ -1,7 +1,9 @@
 package com.practica.ejemplodagger.sis.viewmodel
 
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import android.view.MenuItem
+import android.widget.Toast
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -11,12 +13,15 @@ import com.practica.ejemplodagger.data.domain.EditCategoryUseCase
 import com.practica.ejemplodagger.data.domain.DetailsCategoryUseCase
 import com.practica.ejemplodagger.data.entities.CategoriaEntity
 import com.practica.ejemplodagger.data.repository.CategoriaRepository
+import com.practica.ejemplodagger.data.repository.ProductRepository
+import com.practica.ejemplodagger.sis.ui.view.alerdialog.CategoryIsUseDialog
 import com.practica.ejemplodagger.sis.ui.view.alerdialog.DeleteAlertDialog
 import kotlinx.coroutines.launch
 
 class CategoriaViewModel:ViewModel() {
 
     val repository = CategoriaRepository()
+    val productRepository = ProductRepository()
     var cateogriaList = MutableLiveData<MutableList<CategoriaEntity>>()
     @SuppressLint("StaticFieldLeak")
     val context = MainApplication.appContext
@@ -30,8 +35,13 @@ class CategoriaViewModel:ViewModel() {
 
     fun deleteCategory(category: CategoriaEntity){
         viewModelScope.launch {
-            repository.delete(category)
-            getAllCategorias()
+            if (!isUseCategoria(category.name!!)){
+                repository.delete(category)
+                Toast.makeText(context, "categoria eliminada", Toast.LENGTH_LONG).show()
+                getAllCategorias()
+            }else{
+
+            }
         }
     }
 
@@ -42,6 +52,15 @@ class CategoriaViewModel:ViewModel() {
         }
     }
 
+    suspend fun isUseCategoria(categoria:String): Boolean {
+        val productos = productRepository.getAllProducts()
+        var isUseFlat= false
+        productos.map { producto->
+            if(producto.categoria == categoria) isUseFlat = true
+        }
+        return isUseFlat
+    }
+
     /**atiende la seleccion del context menu de cada item del rv*/
     fun itemSelect(item: MenuItem, category:CategoriaEntity, fragmentManager: FragmentManager){
         when (item.title) {
@@ -50,8 +69,7 @@ class CategoriaViewModel:ViewModel() {
                 detailsCategory.verDetalles(category.image!!, fragmentManager, context!!)
             }
             "Eliminar" -> {
-                val alert = DeleteAlertDialog(category){ deleteCategoria -> deleteCategory(deleteCategoria) }
-                alert.show(fragmentManager, DeleteAlertDialog.TAG)
+                deletecomplete(category,fragmentManager)
             }
             "Editar" -> {
                 val editarCategory = EditCategoryUseCase()
@@ -62,4 +80,16 @@ class CategoriaViewModel:ViewModel() {
     }
 
 
+    fun deletecomplete(category:CategoriaEntity, fragmentManager: FragmentManager){
+        viewModelScope.launch {
+            if(!isUseCategoria(category.name!!)){
+                val alert = DeleteAlertDialog(category){ deleteCategoria -> deleteCategory(deleteCategoria) }
+                alert.show(fragmentManager, DeleteAlertDialog.TAG)
+            }else{
+                val alert = CategoryIsUseDialog()
+                alert.show(fragmentManager, DeleteAlertDialog.TAG)
+            }
+        }
+
+    }
 }
